apps/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = "ETL for Wallet Health"
authors = ["Wallet Health <dev@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.9.3"
requests = "^2.32.3"
alembic = "^1.13.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


